Retixly-2.0-main/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "retixly"
version = "1.0.0"
description = "GUI tool for background removal and image processing"
readme = "README.md"
authors = [{name="RetixlySoft"}]
license = {file = "LICENSE"}
requires-python = ">=3.9"

# dependencies from requirements.txt
dependencies = [
    "PyQt6==6.6.1",
    "Pillow==10.1.0",
    "rembg==2.0.50",
    "numpy==1.26.2",
    "opencv-python==4.8.1.78",
    "boto3==1.34.7",
    "requests==2.31.0",
    "google-auth==2.23.4",
    "google-auth-oauthlib==1.1.0",
    "google-api-python-client==2.108.0",
    "pillow-heif==0.12.0",
    "cryptography",
    "onnxruntime",
]

[project.scripts]
retixly = "main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = [
    "assets/icons/*.png",
    "assets/styles/*.qss",
    "translations/*.qm",
]